# Implement an alternative to `std.concurrency`