--- a/source/geod24/LocalRest.d
+++ b/source/geod24/LocalRest.d
@@ -83,6 +83,7 @@ import vibe.data.json;
 import geod24.concurrency;
 import geod24.LocalRestMessage;
 
+import std.algorithm;
 import std.meta : AliasSeq;
 import std.traits;
 
@@ -342,6 +343,7 @@ public final class RemoteAPI (API) : API
                             break;
 
                         case Message.Type.destoy_pipe_command :
+                            pipeline.close();
                             pipe_terminate = true;
                             break;
 
@@ -422,6 +424,9 @@ public final class RemoteAPI (API) : API
 
     /// Storage of Message Pipeline, Save what has already been created.
     private MessagePipelineRegistry registry;
+
+    /// Collection of all created message pipeline
+    private bool[MessagePipeline] pipeline_collection;
 
     // Vibe.d mandates that method must be @safe
     @safe:
@@ -520,7 +525,7 @@ public final class RemoteAPI (API) : API
 
         public void sleep (Duration d, bool dropMessages = false) @trusted
         {
-            this.childChannel.send(Message(TimeCommand(d, dropMessages)));
+            sendControlMessage(Message(TimeCommand(d, dropMessages)));
         }
 
         /***********************************************************************
@@ -601,7 +606,7 @@ public final class RemoteAPI (API) : API
                 enum mangled = getBestMatch!Overloads;
             }
 
-            this.childChannel.send(Message(FilterAPI(mangled, pretty)));
+            sendControlMessage(Message(FilterAPI(mangled, pretty)));
         }
 
 
@@ -613,7 +618,20 @@ public final class RemoteAPI (API) : API
 
         public void clearFilter () @trusted
         {
-            this.childChannel.send(Message(FilterAPI("")));
+            sendControlMessage(Message(FilterAPI("")));
+        }
+
+
+        /***********************************************************************
+
+            Clear out any filtering set by a call to filter()
+
+        ***********************************************************************/
+
+        private void sendControlMessage (Message msg) @trusted
+        {
+            this.childChannel.send(msg);
+            this.pipeline_collection.keys.filter!(p => !p.isClosed).each!(p => p.consumer.send(msg));
         }
     }
 
@@ -640,11 +658,12 @@ public final class RemoteAPI (API) : API
                             auto pipe = this.registry.locate();
 
                             // If there is no registered message pipeline, or if there is already processing the request.
-                            if ((pipe is null) || ((pipe !is null) && (pipe.isBusy)))
+                            if ((pipe is null) || ((pipe !is null) && ((pipe.isBusy) || (pipe.isClosed))))
                             {
                                 pipe = new MessagePipeline(this.childChannel);
                                 pipe.open();
                                 this.registry.register(pipe);
+                                this.pipeline_collection[pipe] = true;
                             }
 
                             auto msg_req = Message(Command(pipe.getId(), ovrld.mangleof, serialized));