--- a/source/geod24/LocalRest.d
+++ b/source/geod24/LocalRest.d
@@ -637,7 +637,7 @@ public final class RemoteAPI (API) : API
                             auto pipe = new MessagePipeline(this.childChannel);
                             pipe.open();
 
-                            auto msg_req = Message(Command(this.childChannel, pipe.getId(), ovrld.mangleof, serialized));
+                            auto msg_req = Message(Command(pipe.getId(), ovrld.mangleof, serialized));
                             auto msg_res = pipe.query(msg_req, this.timeout);
 
                             if (msg_res.tag == Message.Type.response)