> The existing code is too good.

`LocalRest` is great.

> The purpose of PR 

The purpose of the project was to process data interchangeable between fiber using the same structure as `Golang's Channel`.

> Added Features

We have built-in `timeout` function.
Save the status so that it is safely handled during 'shutdown'.

> Deleted Features

'priority' message type has been removed.

> `LocalRest`

The node is threaded and the request and response are handled by the fiber.