@Geod24 2019-11-06
You need to replace [this import](https://github.com/Geod24/localrest/blob/196aa77d66f2f289fd830a5362aaee1c465542da/source/geod24/LocalRest.d#L83)