@AndrejMitrovic 2019-11-22
Could you write a comment in the PR description / commit message explaining briefly what features this PR implements, and how it's different to std.concurrency? For example what are the benefits / drawbacks compared to std.concurrency?