@MukeunKim 2019-11-22
> Direction of further improvement

PR #26 is still in preparation.
I changed the architecture as well as the 'concurrency'
Applying the experience of PR #25 to #26 will bring you good results.