@MukeunKim 2019-11-22
> > PR #26 is still in preparation.
> 
> So which one needs to be reviewed? #25 or #26? Does one have to be merged before the other?

Please go over #25 first.
I'll prepare more for #26.