@AndrejMitrovic 2019-11-25
> The purpose of the project was to process data interchangeable between fiber using the same structure as Golang's Channel.

Well actually no, the purpose was to implement an alternative to std.concurrency, and Golang channels were just one example of it. 

https://github.com/bpfkorea/agora/issues/308

> @Geod24: We need an alternative way to do message passing, possibly something similar to Go channel.

"Possibly something similar to".

---

I did a test of the performance, and it seems the new code runs the tests slower, and spins the CPUs at 100%:

Before:

```
$ time dub --verror test
All unit tests have been run successfully.
dub --verror test  1.82s user 0.30s system 15% cpu 13.261 total
```

With this PR:

```
~/dev/localrest pull25 $ time dub --verror test
All unit tests have been run successfully.
dub --verror test  40.03s user 0.33s system 207% cpu 19.475 total
```

Is this something that's expected?