@Geod24 2019-11-27
Also, the last 4 commits (named `Fix ...`) should be merged into the second one (named "Developed and applied something similar to `Go Channel`")