@AndrejMitrovic 2019-11-29
> Applied something similar to `Go Channel

So how is it similar? We can create communication channels? The commit message doesn't explain it.