@MukeunKim 2019-11-29
> > Applied something similar to `Go Channel
> 
> So how is it similar? We can create communication channels? The commit message doesn't explain it.

I modified the  commit message as follows. Should I do this?

**Applied the message box similar to `Go Channel`.**
A node runs on a single thread.
And the message is delivered by the Fiber created in each thread.
There are two queues in the `MessageBox`.
One contains a sending message and a fiber.
The other contains receiving messages and fiber.