@AndrejMitrovic 2019-11-29
The new implementation seems to drop messages from the queue which are not matched.

For example:

```D
unittest
{
    import std.concurrency;
    import std.stdio;

    auto process = ()
    {
        size_t message_count = 2;
        while (message_count--)
        {
            receive(
                (int i)     { writefln("Child thread received int: %s", i);
                              ownerTid.send(i); },
                (string s)  { writefln("Child thread received string: %s", s);
                              ownerTid.send(s); }
            );
        }
    };

    auto tid = spawn(process);
    send(tid, 42);
    send(tid, "string");

    // REQUIRED in new API
    Thread.sleep(10.msecs);

    // in new API this will drop all messages from the queue which do not match `string`
    receive(
        (string s)  { writefln("Main thread received string: %s", s); }
    );

    receive(
        (int s)  { writefln("Main thread received int: %s", s); }
    );

    writeln("Done");
}
```

In existing std.concurrency the result is:

```
Child thread received int: 42
Child thread received string: string
Main thread received string: string
Main thread received int: 42
Done
```

In this pull request it's:

```
Child thread received int: 42
Child thread received string: string
Main thread received string: string
Done
```

The `int` message for the main thread was dropped from the message queue. Is this desired? The commit message doesn't say anything about this new behavior.

Also, I can't run the `geod24.concurrency` test-suite to completion. The thread seems to get stuck.