@MukeunKim 2019-11-29
> > I got the following results. Please review it again.
> 
> It fails: https://travis-ci.com/Geod24/localrest/jobs/261509908
> 
> ```
> Running ./localrest-test-library 
> Child thread received int: 42
> Child thread received string: string
> Main thread received string: string
> core.exception.AssertError@source/geod24/concurrency.d(2685): unittest failure
> ```
> 
> The code is here: [033a228](https://github.com/Geod24/localrest/commit/033a228cfd2e3659bc7db2dcb60d5a25261f9488)

Ah! I know the problem now. I'll work it out.